+++ src/hardwareRegisters.h
@@ -67,12 +67,12 @@ extern uint8_t  spi1RxPosition;
 extern uint8_t  spi1TxPosition;
 
 //memory accessors
-unsigned int getHwRegister8(unsigned int address);
-unsigned int getHwRegister16(unsigned int address);
-unsigned int getHwRegister32(unsigned int address);
-void setHwRegister8(unsigned int address, unsigned int value);
-void setHwRegister16(unsigned int address, unsigned int value);
-void setHwRegister32(unsigned int address, unsigned int value);
+uint8_t getHwRegister8(uint32_t address);
+uint16_t getHwRegister16(uint32_t address);
+uint32_t getHwRegister32(uint32_t address);
+void setHwRegister8(uint32_t address, uint8_t value);
+void setHwRegister16(uint32_t address, uint16_t value);
+void setHwRegister32(uint32_t address, uint32_t value);
 
 //timing
 void clk32();//also checks all interrupts
